Spring - handling loggers

+++ src/main/java/pl/kazimierczak/jakub/spring/learn/service/DashboardService.java
@@ -1,6 +1,8 @@
 package pl.kazimierczak.jakub.spring.learn.service;
 
 import org.springframework.stereotype.Service;
+import org.springframework.util.StringUtils;
+import pl.kazimierczak.jakub.spring.learn.api.exception.DashboardException;
 import pl.kazimierczak.jakub.spring.learn.repository.DashboardRepository;
 import pl.kazimierczak.jakub.spring.learn.repository.entity.DashboardEntity;
 import pl.kazimierczak.jakub.spring.learn.service.mapper.DashboardMapper;
@@ -63,12 +65,12 @@ public class DashboardService {//logika biznesowa aplikacji np. czy można wykon
         return mappedDashboardModelList;
     }
 
-    public DashboardModel read(Long id) throws Exception {
+    public DashboardModel read(Long id) throws DashboardException {
         LOGGER.info("read()");
 
         Optional<DashboardEntity> optionalDashboardEntity = dashboardRepository.findById(id);
         DashboardEntity dashboardEntity = optionalDashboardEntity.orElseThrow(
-                () -> new Exception("Unable to find dashboard with id " + id));
+                () -> new DashboardException("Unable to find dashboard with id " + id));
 
         DashboardModel mappedDashboardModel = dashboardMapper.from(dashboardEntity);
 
@@ -77,8 +79,9 @@ public class DashboardService {//logika biznesowa aplikacji np. czy można wykon
         return mappedDashboardModel;
     }
 
-    public void delete(Long id) {
+    public void delete(Long id) throws Exception {
         LOGGER.info("delete(" + id + ")");
+        if (id == null) throw new Exception("Delete id is null");
 
         dashboardRepository.deleteById(id);
